Hey John, i really liked your solution to the rock paper scissors project! Especially the very slick JavaScript code.
The only thing i was missing was some interactivity on the buttons/ images. I added some to it, check it out and use it if you like it.

Check out my rock paper scissors game if you'd like. Happy about any recommendations! (adding mediaqueries and cleaning up the js code are currently on my list)